+++ auditlog/registry.py
@@ -1,8 +1,12 @@
-from typing import Callable, Dict, List, Optional, Tuple
+import copy
+from typing import Any, Callable, Dict, Iterable, List, Optional, Tuple, Union
 
+from django.apps import apps
 from django.db.models import Model
 from django.db.models.base import ModelBase
 from django.db.models.signals import ModelSignal, post_delete, post_save, pre_save
+
+from auditlog.conf import settings
 
 DispatchUID = Tuple[int, str, int]
 
@@ -11,6 +15,8 @@ class AuditlogModelRegistry:
     """
     A registry that keeps track of the models that use Auditlog to track changes.
     """
+
+    DEFAULT_EXCLUDE_MODELS = ("auditlog.LogEntry", "admin.LogEntry")
 
     def __init__(
         self,
@@ -146,6 +152,93 @@ class AuditlogModelRegistry:
         Generate a dispatch_uid.
         """
         return self.__hash__(), model.__qualname__, signal.__hash__()
+
+    def _get_model_classes(self, app_model: str) -> List[ModelBase]:
+        try:
+            try:
+                app_label, model_name = app_model.split(".")
+                return [apps.get_model(app_label, model_name)]
+            except ValueError:
+                return apps.get_app_config(app_model).get_models()
+        except LookupError:
+            return []
+
+    def _get_exclude_models(
+        self, exclude_tracking_models: Iterable[str]
+    ) -> List[ModelBase]:
+        exclude_models = [
+            model
+            for app_model in exclude_tracking_models + self.DEFAULT_EXCLUDE_MODELS
+            for model in self._get_model_classes(app_model)
+        ]
+        return exclude_models
+
+    def _register_models(self, models: Iterable[Union[str, Dict[str, Any]]]) -> None:
+        models = copy.deepcopy(models)
+        for model in models:
+            if isinstance(model, str):
+                for model_class in self._get_model_classes(model):
+                    self.unregister(model_class)
+                    self.register(model_class)
+            elif isinstance(model, dict):
+                model["model"] = self._get_model_classes(model["model"])[0]
+                self.unregister(model["model"])
+                self.register(**model)
+
+    def register_from_settings(self):
+        """
+        Register models from settings variables
+        """
+        if not isinstance(settings.AUDITLOG_INCLUDE_ALL_MODELS, bool):
+            raise TypeError("Setting 'AUDITLOG_INCLUDE_ALL_MODELS' must be a boolean")
+
+        if not isinstance(settings.AUDITLOG_EXCLUDE_TRACKING_MODELS, (list, tuple)):
+            raise TypeError(
+                "Setting 'AUDITLOG_EXCLUDE_TRACKING_MODELS' must be a list or tuple"
+            )
+
+        if (
+            not settings.AUDITLOG_INCLUDE_ALL_MODELS
+            and settings.AUDITLOG_EXCLUDE_TRACKING_MODELS
+        ):
+            raise ValueError(
+                "In order to use setting 'AUDITLOG_EXCLUDE_TRACKING_MODELS', "
+                "setting 'AUDITLOG_INCLUDE_ALL_MODELS' must set to 'True'"
+            )
+
+        if not isinstance(settings.AUDITLOG_INCLUDE_TRACKING_MODELS, (list, tuple)):
+            raise TypeError(
+                "Setting 'AUDITLOG_INCLUDE_TRACKING_MODELS' must be a list or tuple"
+            )
+
+        for item in settings.AUDITLOG_INCLUDE_TRACKING_MODELS:
+            if not isinstance(item, (str, dict)):
+                raise TypeError(
+                    "Setting 'AUDITLOG_INCLUDE_TRACKING_MODELS' items must be str or dict"
+                )
+
+            if isinstance(item, dict):
+                if "model" not in item:
+                    raise ValueError(
+                        "Setting 'AUDITLOG_INCLUDE_TRACKING_MODELS' dict items must contain 'model' key"
+                    )
+                if "." not in item["model"]:
+                    raise ValueError(
+                        "Setting 'AUDITLOG_INCLUDE_TRACKING_MODELS' model must be in the format <app_name>.<model_name>"
+                    )
+
+        if settings.AUDITLOG_INCLUDE_ALL_MODELS:
+            exclude_models = self._get_exclude_models(
+                settings.AUDITLOG_EXCLUDE_TRACKING_MODELS
+            )
+            models = apps.get_models()
+
+            for model in models:
+                if model in exclude_models:
+                    continue
+                self.register(model)
+
+        self._register_models(settings.AUDITLOG_INCLUDE_TRACKING_MODELS)
 
 
 auditlog = AuditlogModelRegistry()